move debug panel to globally available location
